feat(core): ultra-sdk first implementation and testing
This commit adds typedocs implementation on npm script to generate documentation for the project.
Additionally, the commit introduces the first implementation of ᕫltra SDK, which includes testing
for formatter, API calls, and ᕫltra class wrapping the block producer API endpoint. The new ᕫltra
SDK provides a more streamlined and efficient way to interact with the block producer API endpoint
and will improve the overall functionality and user experience of the project.