chore(ultra-sdk): update various files in the Ultra SDK
This commit updates various files in the Ultra SDK related to APIs and utilities. Modifications have
been made to improve the functionality and efficiency of the code. Files related to test cases for
various APIs have also been updated to ensure that all tests are passing. Overall, this commit is
intended to improve the reliability and user experience of the Ultra SDK.