feat(ultra-sdk): added queries function for uniqs (listed, details, owned)
This commit adds four new query functions to the ultra-sdk package, which allows users to interact
with the Ultra blockchain. The new functions include getListedUniqs, which returns a list of all
listed uniques, getUniqDetail, which returns detailed information about a specific unique,
getUniqsOwned, which returns a list of uniques owned by a specific account, and getUosBalance, which
returns the UOS balance of a specific account. These functions will be useful for developers who are
building applications on top of the Ultra blockchain.